+++ src/jvmTest/kotlin/csense/kotlin/extensions/coroutines/CoroutineScopeIOTest.kt
@@ -2,16 +2,13 @@ package csense.kotlin.extensions.coroutines
 
 import csense.kotlin.tests.assertions.*
 import kotlinx.coroutines.*
-import org.junit.*
-import kotlin.coroutines.*
+import org.junit.jupiter.api.*
 
 class CoroutineScopeIOTest {
     @Test
     fun coroutineScopeAsyncIO() = runBlocking {
         val async = asyncIO {
-            val dispatcher = coroutineContext[ContinuationInterceptor]
-                    as CoroutineDispatcher
-            dispatcher.assertAs(Dispatchers.IO)
+            assertDispatcher(Dispatchers.IO)
             "test"
         }
         val value = async.await()
@@ -22,12 +19,23 @@ class CoroutineScopeIOTest {
     fun coroutineScopeLaunchIO() = runBlocking {
         assertCalled { shouldBeCalled ->
             val async = launchIO {
-                val dispatcher = coroutineContext[ContinuationInterceptor]
-                        as CoroutineDispatcher
-                dispatcher.assertAs(Dispatchers.IO)
+                assertDispatcher(Dispatchers.IO)
                 shouldBeCalled()
             }
             async.join()
+        }
+    }
+
+    @Test
+    fun coroutineScopeWithContextIO() = runBlocking {
+        assertCalled { shouldBeCalled ->
+            val async = withContextIO {
+                assertDispatcher(Dispatchers.IO)
+                shouldBeCalled()
+                "test"
+            }
+            async.assert("test")
+
         }
     }
 }